license_server/requirements.txt
# Django License Server Requirements

# Core
Django>=4.2
djangorestframework>=3.14

# Authentication
djangorestframework-simplejwt>=5.3

# CORS
django-cors-headers>=4.3

# Database (optional - for PostgreSQL in production)
# psycopg2-binary>=2.9

# Production server (optional)
# gunicorn>=21.2
# uvicorn>=0.24
